warningの修正

+++ src/Game/Enemy/Barrage/EnemyBarrage.cpp
@@ -15,7 +15,7 @@ EnemyBarrage::EnemyBarrage( Enemy * enemy, int barrageID )
 
 EnemyBarrage::~EnemyBarrage()
 {
-    for( int i = 0; i < this->ebullets.size(); i++ )
+    for( size_t i = 0; i < this->ebullets.size(); i++ )
     {
         if( ebullets[i] != nullptr )
         {
